flask_api/requirements.txt
Flask
Flask-restful
Flask-SQLAlchemy
flask-unittest
flasgger
mysqlclient
python-dotenv
validate-docbr
Faker